gradle: Add FileSetRepository support to Export and Resolve task types
The support is refactored into a convention which is then applied to
the tasks. Tests are update to use this support instead of fixed indexed
repos.
