+++ biz.aQute.bnd.gradle/testresources/exporttask1/runbundles.bndrun
@@ -1,4 +1,3 @@
--standalone: ${.}/index/index.xml
 -runfw: org.apache.felix.framework;version='[5.4.0,5.4.0]'
 -runee: JavaSE-1.7
 -runrequires: osgi.identity;filter:='(osgi.identity=org.apache.felix.eventadmin)'
