+++ biz.aQute.bnd.gradle/testresources/resolvetask1/build.gradle
@@ -8,10 +8,24 @@ plugins {
     id 'biz.aQute.bnd.builder'
 }
 
+repositories {
+    mavenCentral()
+}
+
+configurations {
+  bundles
+}
+
+dependencies {
+    bundles 'org.apache.felix:org.apache.felix.framework:5.4.0'
+    bundles 'org.apache.felix:org.apache.felix.eventadmin:1.4.6'
+}
+
 task create(type: Resolve) {
    description "Resolve ${name}.bndrun"
    group 'export'
    bndrun "${name}.bndrun"
+   bundles = configurations.bundles
 }
 
 task same(type: Resolve) {
@@ -19,6 +33,7 @@ task same(type: Resolve) {
    group 'export'
    bndrun file("${name}.bndrun")
    failOnChanges = true
+   bundles = configurations.bundles
 }
 
 task changefail(type: Resolve) {
@@ -26,10 +41,12 @@ task changefail(type: Resolve) {
    group 'export'
    bndrun { file("${name}.bndrun") }
    failOnChanges = true
+   bundles = configurations.bundles
 }
 
 task resolvefail(type: Resolve) {
    description "Resolve ${name}.bndrun"
    group 'export'
    bndrun { "${name}.bndrun" }
+   bundles = configurations.bundles
 }
